avoid circular dependencies with scope
