+++ packages/system/src/Effect/index.ts
@@ -33,6 +33,7 @@ export * from "./collect"
 export * from "./collectAll"
 export * from "./collectAllUnit"
 export * from "./compose"
+export * from "./core-scope"
 export * from "./core"
 export * from "./delay_"
 export * from "./delay"
