+++ packages/system/src/Effect/race.ts
@@ -3,7 +3,8 @@ import * as E from "../Either"
 import * as Exit from "../Exit/api"
 import * as Fiber from "../Fiber/api"
 import { pipe } from "../Function"
-import { chain, chain_, checkDescriptor, halt, raceWith, succeed } from "./core"
+import { chain, chain_, checkDescriptor, halt, succeed } from "./core"
+import { raceWith } from "./core-scope"
 import { done } from "./done"
 import type { Effect, IO } from "./effect"
 import { map_ } from "./map_"
