+++ packages/system/src/Effect/zipWithPar_.ts
@@ -3,7 +3,8 @@ import type { Exit } from "../Exit/exit"
 import { join } from "../Fiber/api"
 import type { Fiber } from "../Fiber/core"
 import type { FiberID } from "../Fiber/id"
-import { chain_, checkDescriptor, halt, raceWith, transplant } from "./core"
+import { chain_, checkDescriptor, halt } from "./core"
+import { raceWith, transplant } from "./core-scope"
 import type { Effect } from "./effect"
 import { map_ } from "./map_"
