+++ packages/system/src/Fiber/_internal/effect.ts
@@ -1,5 +1,5 @@
 export { asUnit } from "../../Effect/asUnit"
-export { chain_, forkDaemon, halt, succeed, unit } from "../../Effect/core"
+export { chain_, halt, succeed, unit } from "../../Effect/core"
 export { done } from "../../Effect/done"
 export { IO, UIO } from "../../Effect/effect"
 export { fiberId } from "../../Effect/fiberId"
@@ -12,6 +12,7 @@ export { foreach_ } from "../../Effect/foreach_"
 export { map } from "../../Effect/map"
 export { map_ } from "../../Effect/map_"
 export { result } from "../../Effect/result"
+export { forkDaemon } from "../../Effect/core-scope"
 export { tap_ } from "../../Effect/tap_"
 export { zipWithPar_ } from "../../Effect/zipWithPar_"
 export { zipWith_ } from "../../Effect/zipWith_"
