+++ packages/system/src/Schedule/effect.ts
@@ -8,12 +8,12 @@ export {
   chain_,
   effectTotal,
   foldCauseM,
-  forkDaemon,
   provideAll,
   provideAll_,
   succeed,
   unit
 } from "../Effect/core"
+export { forkDaemon } from "../Effect/core-scope"
 export { delay } from "../Effect/delay"
 export { bind, do, let } from "../Effect/do"
 export { done } from "../Effect/done"
